+++ components/cli/src/core/helpers/add-iteractive-and-token-option.ts
@@ -1,4 +1,4 @@
-import { Command, Option } from 'commander';
+import { type Command, Option } from 'commander';
 
 export const addInteractiveAndTokenOption = (command: Command) => {
     command.addOption(new Option('--token_id <token_id>', 'Your access token id.'));